refactor(openai): extract OpenAI-specific utilities into openai/ module
Spec 2.4 complete: Extract OpenAI-specific request/response handling
utilities from provider-utils.ts and request-utils.ts into dedicated
openai/ module.

Create openai/request-handling.ts with:
- appendCsvHeaderValue: CSV header manipulation
- shouldEnableInterleavedThinkingHeader: thinking header checks
- requestWantsReasoningTrace: reasoning trace detection
- sendOpenAiError: OpenAI error response formatting
- summarizeResponsesRequestBody: Responses API body logging

Update imports across codebase:
- provider-strategy/strategies/gemini.ts
- provider-strategy/strategies/standard.ts
- provider-strategy/contexts.ts
- ollama-compat.ts
- routes/responses.ts

Remove moved functions from provider-utils.ts and request-utils.ts.

provider-utils.ts now contains only:
- isRecord, asString: generic type guards
- hasBearerToken: auth header validation
- sendOpenAiError: OpenAI error response (kept for backward compat)

All 579 tests pass.
