+++ src/lib/ollama-compat.ts
@@ -1,4 +1,4 @@
-import { requestWantsReasoningTrace } from "./provider-utils.js";
+import { requestWantsReasoningTrace } from "./openai/index.js";
 
 function isRecord(value: unknown): value is Record<string, unknown> {
   return typeof value === "object" && value !== null;
